some scrollwrite changes
This changes the sound handling so that an in-mission looping sound is played for as long as the text is scrolling, and then stops when the text finishes scrolling.  Right now beep_5.wav is used as a placeholder sound effect; textdraw.wav could also be used, but it would have to be copied into the in-mission sound section in sounds.tbl.

Also adds an important detail to sexp help, and removes trailing whitespace in the script.